fix(cli): extract Session into its own module to break construct import cycle
Every construct imports `Session` from `constructs/project.js`, and
project.js re-exports every construct through the barrel (`index.js`).
That forms an import cycle which only resolved by luck of evaluation
order; under Vite/vitest's module runner it evaluated a subclass before
its base class and threw "Class extends value undefined" (native Node was
unaffected). Earlier changes in this branch perturbed the order and
exposed it.

Move the `Session` class (and its `ConstructExport`, `CheckFilter`,
`SharedFile`, `SharedFileRef` types) into `constructs/session.js`, which
imports construct classes as types only. Construct base classes now import
`Session` from a module that never pulls the barrel, so the cycle is
broken at its root for both Node and bundlers. `Project.__checklyType` is
mirrored as a `PROJECT_CONSTRUCT_TYPE` constant so Session needs no value
import of project.js. Session stays in the public API via the barrel
(`export * from './session.js'`); all internal imports now resolve it from
session.js.
