+++ packages/cli/src/commands/import/plan.ts
@@ -31,7 +31,8 @@ import { generateChecklyConfig } from '../../services/checkly-config-codegen.js'
 import { PackageJsonFile } from '../../services/check-parser/package-files/package-json-file.js'
 import { detectNearestPackageJson, detectPackageManager, knownPackageManagers, PackageManager } from '../../services/check-parser/package-files/package-manager.js'
 import { parseProject } from '../../services/project-parser.js'
-import { ConstructExport, Project, Session } from '../../constructs/project.js'
+import { Project } from '../../constructs/project.js'
+import { ConstructExport, Session } from '../../constructs/session.js'
 import { Diagnostics } from '../../constructs/index.js'
 import { Runtime } from '../../runtimes/index.js'
