+++ packages/cli/src/constructs/__tests__/check-group.spec.ts
@@ -1,7 +1,8 @@
 import { describe, it, expect } from 'vitest'
 
 import { AlertEscalationBuilder, CheckGroup, CheckGroupV1, CheckGroupV2, Diagnostics } from '../index.js'
-import { Project, Session } from '../project.js'
+import { Project } from '../project.js'
+import { Session } from '../session.js'
 import { Bundler } from '../../services/check-parser/bundler.js'
 
 describe('CheckGroup', () => {
