+++ packages/cli/src/constructs/__tests__/private-location.spec.ts
@@ -1,7 +1,8 @@
 import { describe, it, expect } from 'vitest'
 
 import { PrivateLocation, Diagnostics } from '../index.js'
-import { Project, Session } from '../project.js'
+import { Project } from '../project.js'
+import { Session } from '../session.js'
 
 describe('PrivateLocation', () => {
   it('should throw if the same logicalId is used twice', () => {
