+++ packages/cli/src/constructs/alert-channel-subscription.ts
@@ -1,6 +1,6 @@
 import { Ref } from './ref.js'
 import { Construct } from './construct.js'
-import { Session } from './project.js'
+import { Session } from './session.js'
 
 export interface AlertChannelSubscriptionProps {
   alertChannelId: Ref
