+++ packages/cli/src/constructs/api-check.ts
@@ -3,7 +3,7 @@ import fs from 'node:fs/promises'
 import { RuntimeCheck, RuntimeCheckProps } from './check.js'
 import { HttpHeader } from './http-header.js'
 import { BasicAuth, Request } from './api-request.js'
-import { Session, SharedFileRef } from './project.js'
+import { Session, SharedFileRef } from './session.js'
 import { QueryParam } from './query-param.js'
 import { Content, Entrypoint, isContent, isEntrypoint } from './construct.js'
 import { Diagnostics } from './diagnostics.js'
