+++ packages/cli/src/constructs/check-group-v1.ts
@@ -24,7 +24,7 @@ import { Construct } from './construct.js'
 import { MultiStepCheck } from './multi-step-check.js'
 import { PrivateLocationGroupAssignment } from './private-location-group-assignment.js'
 import { Ref } from './ref.js'
-import { Session } from './project.js'
+import { Session } from './session.js'
 import { validateDeprecatedDoubleCheck } from './internal/common-diagnostics.js'
 import { CheckRetryStrategy } from './check.js'
 import { MonitorRetryStrategy } from './monitor.js'
