+++ packages/cli/src/constructs/check.ts
@@ -4,7 +4,7 @@ import { Construct } from './construct.js'
 import { AlertChannel, AlertChannelRef } from './alert-channel.js'
 import { EnvironmentVariable } from './environment-variable.js'
 import { AlertChannelSubscription } from './alert-channel-subscription.js'
-import { Session } from './project.js'
+import { Session } from './session.js'
 import type { Region } from '../index.js'
 import type { CheckGroupV1, CheckGroupV2, CheckGroupRef } from './check-group.js'
 import { PrivateLocation, PrivateLocationRef } from './private-location.js'
