+++ packages/cli/src/constructs/heartbeat-monitor.ts
@@ -1,5 +1,5 @@
 import { Monitor, MonitorProps } from './monitor.js'
-import { Session } from './project.js'
+import { Session } from './session.js'
 import { DateTime } from 'luxon'
 import { CheckTypes } from '../constants.js'
