+++ packages/cli/src/constructs/multi-step-check.ts
@@ -1,7 +1,7 @@
 import fs from 'node:fs/promises'
 
 import { CheckProps, RuntimeCheck, RuntimeCheckProps } from './check.js'
-import { Session, SharedFileRef } from './project.js'
+import { Session, SharedFileRef } from './session.js'
 import { Content, Entrypoint, isContent, isEntrypoint } from './construct.js'
 import { CheckTypes } from '../constants.js'
 import { PlaywrightConfig } from './playwright-config.js'
