+++ packages/cli/src/constructs/opsgenie-alert-channel.ts
@@ -1,5 +1,5 @@
 import { AlertChannel, AlertChannelProps } from './alert-channel.js'
-import { Session } from './project.js'
+import { Session } from './session.js'
 
 export type OpsgeniePriority = 'P1' | 'P2' | 'P3' | 'P4' | 'P5'
