+++ packages/cli/src/constructs/private-location-check-assignment.ts
@@ -1,6 +1,6 @@
 import { Ref } from './ref.js'
 import { Construct } from './construct.js'
-import { Session } from './project.js'
+import { Session } from './session.js'
 
 export interface PrivateLocationCheckAssignmentProps {
   privateLocationId: Ref
