+++ packages/cli/src/constructs/url-monitor.ts
@@ -1,7 +1,7 @@
 import { Diagnostics } from './diagnostics.js'
 import { validateResponseTimes } from './internal/common-diagnostics.js'
 import { Monitor, MonitorProps } from './monitor.js'
-import { Session } from './project.js'
+import { Session } from './session.js'
 import { UrlRequest } from './url-request.js'
 
 /**
